% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{append_wds_unit_classes}
\alias{append_wds_unit_classes}
\title{Append Unit-Classes to the Test Case's Working Data Set}
\usage{
append_wds_unit_classes(tc)
}
\arguments{
\item{tc}{a \code{openNCA_testcase} object}
}
\value{
Returns the \code{openNCA_testcase} object \code{tc} with the
appropriate units, which can be found in the MCT, attached to the
\code{unit_class} slot
}
\description{
Append Unit-Classes to the Test Case's Working Data Set
}
