% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{append_wds_units}
\alias{append_wds_units}
\title{Append Units to the Test Case's Working Data Set}
\usage{
append_wds_units(tc)
}
\arguments{
\item{tc}{a \code{openNCA_testcase} object}
}
\value{
Returns the \code{openNCA_testcase} object \code{tc} with the
appropriate units, which can be found in the MCT, attached to the
\code{units} slot
}
\description{
Append Units to the Test Case's Working Data Set
}
