% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{assign_wds_labels}
\alias{assign_wds_labels}
\title{Assign Working Data Set Labels}
\usage{
assign_wds_labels(tc)
}
\arguments{
\item{tc}{\code{openNCA_testcase} object, the input test case}
}
\value{
The input, \code{tc}, with labeled assigned to the Working Data Set
(\code{WDS} slot)
}
\description{
This function uses the information provided in the MCT of the input test case
to first match parameters in the working data set to set conventions using
RegEx, these parameters are then assigned appropriate labels and units, which
are stored in each parameter's "label" attribute - this label will be picked
up by {gtsummary} further down the pipeline
}
