% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{filter_wds_exclusions}
\alias{filter_wds_exclusions}
\title{Remove Profiles Flagged for Exclusion from the Working Data Set}
\usage{
filter_wds_exclusions(tc, flg, profile, by)
}
\arguments{
\item{tc}{\code{openNCA_testcase} object, the input test case}

\item{flg}{\code{chr}, the parameter name in the WDS which corresponds to the
exclusion flag}

\item{profile}{\code{chr}, the parameter name in the WDS which corresponds to
an identifier of one experimental unit of analysis}

\item{by}{\code{chr}, the parameter name in the WDS which corresponds to
summary grouping parameter}
}
\value{
tc \code{openNCA_testcase} object, the input test case with the
modified WDS post exclusion
}
\description{
Remove Profiles Flagged for Exclusion from the Working Data Set
}
