% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{get_ard}
\alias{get_ard}
\title{Get ARD Dataset from Test-Case}
\usage{
get_ard(tc)
}
\arguments{
\item{tc}{an \code{openNCA_testcase} object}
}
\value{
The ARD, a \code{data.frame} for the test case
}
\description{
Simple convenience function for accessing the ARD Dataset
}
