% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{get_flg}
\alias{get_flg}
\title{Get FLG Dataset from Test-Case}
\usage{
get_flg(tc)
}
\arguments{
\item{tc}{an \code{openNCA_testcase} object}
}
\value{
The FLG, a \code{data.frame} for the test case
}
\description{
Simple convenience function for accessing the FLG Dataset
}
