% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{get_mct}
\alias{get_mct}
\title{Get MCT Dataset from Test-Case}
\usage{
get_mct(tc)
}
\arguments{
\item{tc}{an \code{openNCA_testcase} object}
}
\value{
The MCT, a \code{data.frame} for the test case
}
\description{
Simple convenience function for accessing the MCT Dataset
}
