% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{get_param}
\alias{get_param}
\title{Get PARAM Dataset from Test-Case}
\usage{
get_param(tc)
}
\arguments{
\item{tc}{an \code{openNCA_testcase} object}
}
\value{
The PARAM Dataset, a \code{data.frame} for the test case
}
\description{
Simple convenience function for accessing the PARAM Dataset
}
