% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{get_parameter_label}
\alias{get_parameter_label}
\title{Get Label for Parameter from NCA Dependency List}
\usage{
get_parameter_label(x)
}
\arguments{
\item{x}{\code{chr}, the parameter name}
}
\value{
\code{chr} the parameter label
}
\description{
Get Label for Parameter from NCA Dependency List
}
\examples{
\dontrun{get_parameter_label("AUCINFP")}
}
