% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{get_parameter_unit}
\alias{get_parameter_unit}
\title{Get Unit for Parameter from NCA Dependency List}
\usage{
get_parameter_unit(x, mct)
}
\arguments{
\item{x}{\code{chr}, the parameter name}

\item{mct}{\code{data.frame}, the MCT data set from the test case}
}
\value{
\code{chr} the parameter label
}
\description{
Get Unit for Parameter from NCA Dependency List
}
