% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{get_parameter_unit_class}
\alias{get_parameter_unit_class}
\title{Get Unit Class for Parameter from NCA Dependency List}
\usage{
get_parameter_unit_class(x)
}
\arguments{
\item{x}{\code{chr}, the parameter name}
}
\value{
\code{chr} the parameter label
}
\description{
Get Unit Class for Parameter from NCA Dependency List
}
