% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{get_wds}
\alias{get_wds}
\title{Get WDS from Test-Case}
\usage{
get_wds(tc)
}
\arguments{
\item{tc}{an \code{openNCA_testcase} object}
}
\value{
The working dataset, a \code{data.frame} for the test case
}
\description{
Simple convenience function for accessing the Working Dataset
}
