% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{get_wds_vars}
\alias{get_wds_vars}
\title{Get Working Data Set Parameter Names}
\usage{
get_wds_vars(tc)
}
\arguments{
\item{tc}{\code{openNCA_testcase} object, the input test case}
}
\value{
the current set (names) of parameters in the WDS
}
\description{
Get Working Data Set Parameter Names
}
