% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{gm_cv}
\alias{gm_cv}
\title{Geometric Coefficient of Variation}
\usage{
gm_cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the geometric coefficient of variation of \code{x}, a single numeric
}
\description{
Geometric Coefficient of Variation
}
\examples{
\dontrun{gm_cv(1:10)}
}
