% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric Mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the geometric mean of \code{x}, a single numeric
}
\description{
Geometric Mean
}
\examples{
\dontrun{gm_mean(1:10)}
}
