% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{little_n}
\alias{little_n}
\title{Compute "Little n" for the exclusion profile}
\usage{
little_n(x)
}
\arguments{
\item{x}{a \code{lgl} vector with TRUE indicating an inclusion of that
experimental unit from the summaries}
}
\value{
the value for little n, which is the number of experimental units
considered in the summaries (i.e. the total number minus exclusions)
}
\description{
Compute "Little n" for the exclusion profile
}
