% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{load_test_case}
\alias{load_test_case}
\title{Load NCA Test Case from Path}
\usage{
load_test_case(
  path = NULL,
  ard_path = NULL,
  flg_path = NULL,
  mct_path = NULL,
  param_path = NULL
)
}
\arguments{
\item{path}{directory containing the ARD, FLG, MCT, and PARAM data sets in
csv format}

\item{ard_path}{path to the ARD in csv format}

\item{flg_path}{path to the FLG in csv format}

\item{mct_path}{path to the MCT in csv format}

\item{param_path}{path to the PARAM in csv format}
}
\value{
a \code{openNCA_testcase} object
}
\description{
The function works in one of two ways, either the user specifies a path to a
directory containing a collection of four csv data sets:
\itemize{
\item The Model Configuration Template (MTC)
\item The Analysis Ready Dataset (ARD)
\item The Flags data set (FLG)
\item The PK Parameters data set (PARAM)
Each identified with the patterns:  "ARD", "MCT", "FLG, and "PARAM" in the
file names. Or the user has the options to include paths to each file.
}
}
\examples{
\dontrun{tc <- load_test_case("path/to/testcase")}
}
