% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{lower_range}
\alias{lower_range}
\title{Lower Range}
\usage{
lower_range(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the lower range of \code{x}
}
\description{
Lower Range
}
\examples{
\dontrun{lower_range(1:10)}
}
