% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_blank}
\alias{make_blank}
\title{Return Blank Factor}
\usage{
make_blank(x)
}
\arguments{
\item{x}{any object}
}
\value{
always returns a blank factor (length 1)
}
\description{
This function takes any object and returns a blank factor
for use as the .else argument in map_if
}
