% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{make_test_case}
\alias{make_test_case}
\title{Create an openNCA Test Case Object}
\usage{
make_test_case(ard, flg, mct, param)
}
\arguments{
\item{ard}{\code{data.frame}}

\item{flg}{\code{data.frame}}

\item{mct}{\code{data.frame}}

\item{param}{\code{data.frame}}
}
\value{
a \code{openNCA_testcase} object
}
\description{
This function creates an object of class \code{openNCA_testcase} which
represents a unit of analysis. At a minimum this object contains four data
sets; The Analysis Ready Data set (ARD), the Model Configuration Template
data set (MCT), the Flags data set (FLG), and the parameters data set
(PARAM). These are all stored as \code{data.frame} objects.
}
