% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{mean_pl_sd}
\alias{mean_pl_sd}
\title{Mean Plus One Standard Deviation}
\usage{
mean_pl_sd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the arithmetic mean of \code{x} plus one standard deviation, a single numeric
}
\description{
Mean Plus One Standard Deviation
}
\examples{
\dontrun{mean_pl_sd(1:10)}
}
