% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nca_dependency_list}
\alias{nca_dependency_list}
\title{NCA dependency list}
\format{
A list with 204 slots, one per parameter, each slow has a further 8 nested slots

\describe{
\item{parameter label}{price, in US dollars}
\item{cdisc_label}{}
\item{regex}{A RegEx pattern that links vars in the params data to their slot here}
\item{valid_models}{}
\item{display_list_models}{}
\item{predecessors}{}
}
}
\usage{
nca_dependency_list
}
\description{
A list containing attributes for NCA PK parameters
}
\keyword{datasets}
