% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_factory}
\alias{pred_factory}
\title{Predicate Function Factory}
\usage{
pred_factory(set)
}
\arguments{
\item{set}{a \code{chr} vector representing the set of allowed values}
}
\value{
This function will return a function that can be used as a predicate
in purrr::map_if
}
\description{
This 'function factory' will make a predicate for if a variable is in the set
of outputunits
}
\examples{
\dontrun{letters_pred <- pred_factory(letters)}
}
