% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_labels.R
\name{select_wds_pars}
\alias{select_wds_pars}
\title{Select Working Data Set Parameters}
\usage{
select_wds_pars(tc, ...)
}
\arguments{
\item{tc}{\code{openNCA_testcase} object, the input test case}

\item{...}{parameter names (as symbols ala {dplyr}) to keep in the working
data set, any parameters not included are removed.}
}
\value{
\code{tc} \code{openNCA_testcase} object, the input test case with
the modified WDS post parameter selection.
}
\description{
From the working data set, take these parameters for display in summary
tables
}
\examples{
\dontrun{select_wds_pars(tc, SDEID, AUCLAST, TMAX1)}
}
