% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{update_label}
\alias{update_label}
\title{Update Object's Label Attribute}
\usage{
update_label(x, label)
}
\arguments{
\item{x}{any object}

\item{label}{the new label}
}
\value{
the input, \code{x}, with updated label
}
\description{
Update Object's Label Attribute
}
