% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{update_label_df}
\alias{update_label_df}
\title{Update \code{data.frame} Variable Lables}
\usage{
update_label_df(df, ...)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{...}{The variables to label, should be called like: \code{var_1 =
"label_1"}}
}
\value{
The input, \code{df}, with the listed labels applied
}
\description{
This function will update multiple variable labels at once
}
\examples{
\dontrun{
  update_label_df(iris, Sepal.Length = "Sepal Length",
                  Species = "Iris Species")
                  }
}
