% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{update_label_df_var}
\alias{update_label_df_var}
\title{Update the Label of a Variable in a \code{data.frame}}
\usage{
update_label_df_var(df, var, label)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{var}{\code{chr}, the name of the variable to add a label to}

\item{label}{\code{chr}, the new label to give to \code{var}}
}
\value{
the input \code{data.frame} \code{df} with the variable \code{var}
updated with a label
}
\description{
Update the Label of a Variable in a \code{data.frame}
}
\examples{
 \dontrun{update_label_df_var(iris,
                                        "Sepal.Lenght", 
                                        "Sepal Length (cm)")}
}
