% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fns.R
\name{upper_range}
\alias{upper_range}
\title{Upper Range}
\usage{
upper_range(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
the upper range of \code{x}
}
\description{
Upper Range
}
\examples{
\dontrun{upper_range(1:10)}
}
