% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_case.R
\name{load_test_case}
\alias{load_test_case}
\title{Load NCA Test Case from Path}
\usage{
load_test_case(path)
}
\arguments{
\item{path}{directory containing the ARD, FLG, MCT, and PARAM datasets in csv
format}
}
\value{
a \code{openNCA_testcase} object
}
\description{
Load NCA Test Case from Path
}
